Try struct based lists
IMO this list structure is not as good as the prefixed type, as it
doesn't realy work as a function parameter, etc.

I think I will do a more prefix-based type, but keep the object
system scrapped. IMO the prefixes are only really useful in the
context of data structures, so it will be better implemented
just for those, and without all the additional computational
overhead.