Get basic texture loading working
This displays a smiley face in the browser and on PC.

TODO: Maybe use highp float on PC and mediump in browser? Although
then that would be inconsistent...

Or could just use highp float everywhere, but that might not be
great for mobile devices... although perhaps that doesn't matter.

Or I suppose it could be a project setting.