Actually update alloc_last_pool This significantly increases the speed in a scenario where nodes are being linearly allocated.
We could always go back to the recursive tree allocator I implemented
a while ago for even better performance in some of these cases...

Or perhaps we could have a linked list of pools with empty masks?
That way we wouldn't need to iterate through any that were full. That
would probably be a good optimization.