Get node creation/destruction working for the most part
However, benchmarks indicate that essentially all of the overhead
is currently in the allocator structure. This is despite the small
goal to have a very fast allocator structure. Very sad...

I think a linked list structure will definitely be better.

In fact, a good linked list could be substantially better than the
pool structure.

Consider: two lists. The free list and the non-free list. When new
nodes are added to the free list, we allocate a block of memory and
add everything to the free list.

Impact on the node size: need two pointers for the doubly-linked list.
However, we don't need the pool pointer. So it's a net increase of
8 bytes.

Not the best ever, but has two benefits:
- Basically as fast as possible to iterate over "all nodes of a given
  type"
- No changing overhead in allocation or deallocation. Should stay fast
  no matter how many nodes there are.

The hope is the pointer chasing is not a cache problem, because memory
is allocated in blocks (that are never freed), and the pointers just
point around those blocks (even though they do make the nodes bigger,
which is a bit of a disadvantage for cache friendliness).