Better node creation benchmark.
The new benchmark tests, essentially, out-of-order creation,
with a seemingly reasonable sized load (16 million nodes).

In this case, the best-case node creation is still 0.1 us. Also,
best-case node deletion still seems to be 0.05 us.

However, the out-of-order node deletion, (going through the array
in 256 passes, incremeing by 256 within each pass, in this case)
has a time of 0.4 us.

Then, the new node creation, which in theory is re-using the old
allocated nodes but now in a much more shuffled fashion, takes 0.3 us.

It is possible of course that it would be even slower in a worse case
scenario.

However, there are a lot of reasonable games that won't use close to
16 million nodes. So I think this shows that the node creation and
deletion is pretty reasonable at this point in time.