Potentially fix set_gpos; New transform tests.
I have written some more tests for the transform logic, because that
stuff needs to be solid. This led to fixes for set_gpos.

It is a little suspicious IMO that all the tests are passing despite
using exact floating point math. But, it does seem like the
global position is being updated quite substantially and still
working, so I suppose it is fine..?