Cleanup transform API
This fixes one strange warning where GCC seemed to think I was
copying too much data into the transform union.

I don't know if that was a real warning or just something got messed
up.

Either way the API is now cleaner in my opinion. This should also
possibly make the node type smaller.

In any case, I do need to change the linked lists slightly because
they are starting to build up a lot of data they don't need, and
although that's probably not a huge performance concern, it does
make the initializers for the lists pretty weird to write out
in code.