Stress test renderer to some degree, measure frame time.
So it seems like the renderer can probably handle 10,000 sprites
on my computer but not many more.

I'm guessing most of the problem comes from all the sorting. I wonder
if it would be faster if each render command was smaller.

Also perhaps the opaque list could simply not be sorted if it gets
too large? Or could be sorted only partially, somehow?

Things to think about.