Working resource building... This is still very awkward, as 'scan' and 'build-res' really need to be combined into a single command, or something.
But, it does work. It is possible to build a game that reparents
a StaticSprite to the root, and draws a texture generated by
pony.

Note that I was flipping the V coordinates of the UV incorrectly.
This commit fixes that. Basically, they were being flipped twice,
I think, when they only need to be flipped once to be correct.