Start to consolidate different features
I might want to de-consolidate these again in the future, as it
may be useful to e.g. regenerate the build file without also
redoing all the texture packing, and especially without regenerating
the res header file, as that causes a lot of rebuilding.

Also, it may be worth generating the header file, then seeing if
it has changed at all from the previous header file, and simply not
rewriting the header if there was no change.

That way, changes that only affected the resource loading, and not
the file structure, would not require a code rebuild.